BashScriptforGatheringandStoringNetworkIps
This Bash Script was created to simply the ipsweeping process and to provide the user with a cleaner output that is much easier to read. With such a long command required for every ip address preforming this task manually runs the risk of error therefore it is more effective to use the script.